Moved locations task from commtrack to locations.
